backend/requirements.txt
# Production dependencies for Resume Scanner Backend
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0

# AI/ML
openai==1.3.7
tiktoken==0.5.2

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Database (optional)
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring (optional)
prometheus-client==0.19.0
structlog==23.2.0